requirements.txt
llvmlite==0.36.0
numba==0.53.1
numpy==1.20.3
Pillow==8.2.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
